Add files via upload
Created with Colab.
This notebook has overhauled the structure of the output of the LLM. 
The output now works in 2 steps to go from XML to JSON. 
The LLM needs to follow the codebook chronologically for each variable, and use the established variable codes. 
In addition, a new column is created by the LLM for historical detail/ descriptive strings/ context. 
The notebook uses Claude 4 Sonnet. The temperature is set to 0,2.